Fix) boardNo가 null일때  발생하는 오류 해결 - long - BigInteger로 바꿈 Update) 글 클릭 및 상세 확인 시 Hit 수 증가 기능 추가

+++ src/main/java/com/cdg/ultraViolet/domain/Board.java
@@ -1,5 +1,6 @@
 package com.cdg.ultraViolet.domain;
 
+import java.math.BigInteger;
 import java.sql.Date;
 
 import lombok.Data;
@@ -7,7 +8,7 @@ import lombok.Data;
 @Data
 public class Board {
 
-	private long boardNo;
+	private BigInteger boardNo;
 	private String contents;
 	private String title;
 	private int userNo;
